put mesh-related auxiliary functions into a file mesh.c
