+++ Data/JMM3dial_ibox_slog.txt
@@ -1,9 +1,9 @@
 17	 2.4235e-05	 8.1882e-06	 2.2483e-05	1.1447e-05	2.8131e-03	6.0379e-04	0.000489	2.633	2.616
-33	 3.1815e-06	 1.0251e-06	 2.9514e-06	1.4414e-06	1.0785e-03	1.6474e-04	0.00171	2.755	2.747
-65	 5.7895e-07	 1.8250e-07	 5.3708e-07	2.5739e-07	3.8379e-04	4.7693e-05	0.006552	2.805	2.801
-129	 1.2713e-07	 4.3125e-08	 1.1794e-07	6.0912e-08	1.4036e-04	1.4231e-05	0.024281	2.840	2.838
-257	 3.6376e-08	 1.0743e-08	 3.3745e-08	1.5186e-08	4.9557e-05	4.2583e-06	0.096541	2.863	2.862
-513	 9.9054e-09	 2.6943e-09	 9.1892e-09	3.8098e-09	1.7639e-05	1.2753e-06	0.376268	2.872	2.871
-1025	 2.6162e-09	 6.7596e-10	 2.4270e-09	9.5601e-10	6.2860e-06	3.8143e-07	1.51757	2.875	2.875
-2049	 6.7796e-10	 1.6953e-10	 6.2893e-10	2.3979e-10	9.4070e-06	1.1408e-07	6.19241	2.877	2.877
-4097	 1.7362e-10	 4.2490e-11	 1.6106e-10	6.0103e-11	2.8315e-06	3.4515e-08	25.8691	2.879	2.879
+33	 3.1815e-06	 1.0251e-06	 2.9514e-06	1.4414e-06	1.0785e-03	1.6474e-04	0.001662	2.755	2.747
+65	 5.7895e-07	 1.8250e-07	 5.3708e-07	2.5739e-07	3.8379e-04	4.7693e-05	0.006368	2.805	2.801
+129	 1.2713e-07	 4.3125e-08	 1.1794e-07	6.0912e-08	1.4036e-04	1.4231e-05	0.02427	2.840	2.838
+257	 3.6376e-08	 1.0743e-08	 3.3745e-08	1.5186e-08	4.9557e-05	4.2583e-06	0.095675	2.863	2.862
+513	 9.9054e-09	 2.6943e-09	 9.1892e-09	3.8098e-09	1.7639e-05	1.2753e-06	0.370199	2.872	2.871
+1025	 2.6162e-09	 6.7596e-10	 2.4270e-09	9.5601e-10	6.2860e-06	3.8143e-07	1.50091	2.875	2.875
+2049	 6.7796e-10	 1.6953e-10	 6.2893e-10	2.3979e-10	9.4070e-06	1.1408e-07	6.25246	2.877	2.877
+4097	 1.7362e-10	 4.2490e-11	 1.6106e-10	6.0103e-11	2.8315e-06	3.4515e-08	26.8664	2.879	2.879
